render a single whitespace for soft break
because it is the correct behavior in terms of text search.

For example,

```
foo
bar
```

is rendered as

```
foo bar
```

on the preview. However, searching text with 'foo bar' is not hit
because the internal text is rendered as `"foo\nbar"` for now. By this
commit, the text search works fine since the internal text becomes
`"foo bar"`.