directly get window inner size from `WKWebView`'s frame on macOS
because `tao::Window::inner_size` does not work. It obtains
size of `contentView`'s frame. However `wry::WebView` replaces the
content view of the `NSWindow` with their own `NSView`. As the result
the replaced view is no longer updated on resizing the window but
`tao::Window` still refers the original view.